♻️ Refactor option, menu, result
